Merge ndm and pte data map to enable cases like lora, where we introduce
- shared weights in the pte file (this is true of any xnnpack pte file, as weights are placed in shared space by default now).
- shared weights in external file

Pass the merged data map into method, so it can access both.

Note: have to remove the const qualifiers, as the flat tensor data map is modified.

Differential Revision: [D76447708](https://our.internmc.facebook.com/intern/diff/D76447708/)

[ghstack-poisoned]

+++ extension/flat_tensor/flat_tensor_data_map.cpp
@@ -205,7 +205,9 @@ ET_NODISCARD Error FlatTensorDataMap::load_data_into(
 }
 
 ET_NODISCARD Result<uint32_t> FlatTensorDataMap::get_num_keys() const {
-  return key_to_map_index_.size();
+  // Guaranteed safe, as the segment_index is a uint32_t, which means
+  // that there can't be more than uint32_t keys.
+  return static_cast<uint32_t>(key_to_map_index_.size());
 }
 
 ET_NODISCARD Result<const char*> FlatTensorDataMap::get_key(
@@ -246,7 +248,7 @@ ET_NODISCARD Error FlatTensorDataMap::merge(const NamedDataMap* other) {
   // Place keys into the map.
   for (uint32_t i = 0; i < num_keys; i++) {
     const char* key = other->get_key(i).get();
-    key_to_map_index_[key] = merged_maps_.size();
+    key_to_map_index_[key] = static_cast<int64_t>(merged_maps_.size());
   }
 
   merged_maps_.push_back(other);
@@ -322,7 +324,7 @@ ET_NODISCARD Error FlatTensorDataMap::merge(const NamedDataMap* other) {
       "FlatTensor segments is nullptr, malformed PTD file.");
 
   // Add keys to the map.
-  std::unordered_map<std::string, int32_t> key_to_map_index;
+  std::unordered_map<std::string, int64_t> key_to_map_index;
   for (int i = 0; i < flat_tensor->named_data()->size(); i++) {
     const auto* named_data = flat_tensor->named_data()->Get(i);
     key_to_map_index[named_data->key()->c_str()] = -1;
